requirementsfork8.txt
gradio==5.9.1
python-dotenv>=1.0.0
pydantic==2.6.4
fastapi==0.110.0
databricks-vectorsearch>=0.40
langchain>=0.2.0
langchain-community>=0.2.0
mlflow>=2.12.0
databricks-sdk>=0.28.0
requests>=2.31.0
